+++ setup.py
@@ -23,7 +23,8 @@ setup(
     data_files=[
         ('share/man/man1', ['man/en/sos-report.1', 'man/en/sos.1',
                             'man/en/sos-collect.1', 'man/en/sos-clean.1',
-                            'man/en/sos-mask.1', 'man/en/sos-help.1']),
+                            'man/en/sos-mask.1', 'man/en/sos-help.1',
+                            'man/en/sos-upload.1']),
         ('share/man/man5', ['man/en/sos.conf.5']),
         ('share/licenses/doca-sosreport', ['LICENSE']),
         ('share/doc/doca-sosreport', ['AUTHORS', 'README.md']),
